fixed minor msvc8 warnings
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@2119 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxquery.cpp
@@ -1018,7 +1018,7 @@ private:
 		if ( m_iBits==0 && m_uMask!=0 )
 		{
 			for ( uint64_t dMask = m_uMask; dMask; dMask >>=1 )
-				m_iBits += dMask & 1;
+				m_iBits += (int)( dMask & 1 );
 		}
 
 		// current working formula is num_nodes^2 * num_hits
